requirements.txt
Appium-Python-Client==0.47
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
configparser==3.8.1
ddt==1.2.1
idna==2.8
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.2
pytest==5.1.1
pytest-base-url==1.4.1
pytest-check==0.3.5
pytest-dotenv==0.4.0
pytest-html==1.22.0
pytest-metadata==1.8.0
pytest-selenium==1.17.0
pytest-variables==1.9.0
python-dotenv==0.10.3
PyYAML==5.1.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
xlrd==1.2.0
XlsxWriter==1.1.9
xlutils==2.0.0
xlwt==1.3.0
zipp==0.6.0
